I am ready to resume. Let me erase all that I did today.
What I will do is by replicating the eventually monad in Spiral and then make it work on some toy examples elsewhere. The I will take the chance to redo the whole library. It won't be particularly difficult.

Let me start.

Regarding the evetually monad, I remember playing a lot with it in late 2016 before I threw in the towel.