``` System.Console.WriteLine("I am in a.") System.Console.WriteLine("[state=, 123]") System.Console.WriteLine("I am in a.") System.Console.WriteLine("[state=, 456]") System.Console.WriteLine("I am in b.") System.Console.WriteLine("[state=, 123]") System.Console.WriteLine("I am in b.") System.Console.WriteLine("[state=, 456]") System.Console.WriteLine("I am in c.") System.Console.WriteLine("[state=, 123]") ```
This is so good. This works exactly as I'd envisioned it.

I will admit the interface for the functions is a bit different though, but I should just work with this.

Actually, I will resist the urge to parallelize literally everything. I will stick to just the layers for this. Otherwise this will grind the compilation to a halt and I do not need to make everything concurrent. The benefits will only go up to a certain point anyway.