"9:45am. I am up. I am still obsessed. Right now I am trying to solve the code bloat issue.
With a straightforward pass, I'd get O(3) compilation one for layers with state and one for without. With wavefront I'd get O(2*(3n-2))compilation in the number of layers.

It just makes the number of states explode. Though thankfully the number of layers in an RNN is usually not that big. Some of the usual solution like boxing would not work because everything needs to be in fact CPS'd due to memory allocations.

11:25am. I am just thinking and thinking and thinking...

I was going to read HxH when I got up, but then I got obsessed again over solving this. I've done nothing but pace around the room trying to untangle the knots.

With the regards to the code bloat issue I have a confession to make - I fucked up.

The place where I fucked up would be my CPS'd kind of memory allocation. It seemed a good idea at the time, but now it is starting to stand out as the main reason why I won't be able to fix the code bloat issue.

Strictly speaking compiling the network for 100 different states is not a problem. The problem right now is that each of the states are enormous.

Because memory allocation makes all the calls non-tail recursive I have absolutely no chance of using the tricks I showed in the parser segment of the tutorial to combat it.

Ahhhhh, I fucked up pure and simple. Not only does the memory management scheme make the code much more complicated, but now it literally makes the recurrent segment non-viable.

11:40am. Note to self, do not use CPS for resource management in Spiral ever again. It should be used for control flow and that only.

11:45am. So, have I converged to the next step? I think I have. Before doing anything else, I am going to have to rewrite the library in order to support proper region based allocation.

This will take me a few days. Obviously, I will not be finishing the char RNN just yet. I will add support for allocating streams and events in regions while I am at it.

11:50am. Let me take a break here. This is so strenuous despite me not lifting a finger."