I need to think about it a little.
Right now an idea occurs to me. In tandem with the idea of splitting the forward pass's execution from allocation, I could also use that to calculate region sizes before running them.

The way I am doing MM right now is not necessarily very memory efficient. The thing I have in mind is much better, but I am not sure I want to go for it since I'll have to suffer a loss of flexibility. But that kind of flexibility is not something I need right now.

Ah...I need to think about it. Let me do the other stuff and I will try programming when I get back.