https://www.youtube.com/watch?v=IUJvnIamMo8&list=PL0oekSefhQVJdk0hSRu6sZ2teWM740NtL&index=48
I'll stop at this point for the day. I'll leave virtual memory and memory allocation for tomorrow.

I learned a few new things from watching all of that, but by large I was familiar with most of the material which is why I could go through it is such quantity.

Tomorrow I will actually be able to start what I intended to do today. I could have started off today with memory allocation, but I wanted to see what the course had to offer.

Leaving the course aside, I am not really pleased with how much I could scrunge up on memory management on the Internet. I am going to have to look more carefully.

Well, this is a deep topic so I will give it some time. I won't aim to finish it all tomorrow. I am going to go through it in depth. Never again will I be hobbled by the lack of memory management ability.

Also, I've made a decision. I am going to change the semantics of layout types in order to simplify them. I am going to extend them to work on every datatype, and in turn in order to make this easier, I am going to unpack them as aggressively as union types. I am not going to bother repacking them in various module function as well as it is an annoyance to keep track of that.

This is practically minded...I haven't used them even once since their invention for anything of value, but going forward I am going to have to use them to preserve information from join points for layer combinators and I need them for that purpose.

...Actually, thinking on it more I have used them for things of value, but that is to keep the naked types from spilling on the term level. There are better way of using layout types.

Spiral is due for a redesign. This is good. This is fundamentally how the language should evolve - in response to feedback from my own coding. Maybe this is the true way after all?