I've fixed the type errors just now, I'll fix the actual errors tomorrow. No doubt the language is in a broken state after the redesign.
Right now I am too tired. I did well as usual. Tomorrow I will deal with this and then start work on the new GPU allocator.

It won't be hard at all to do. It was more difficult to decide how I actually wanted to do it. Based on how things are going, it is unlikely that Racket is the solution to my GC woes.

It will like be a good long while until I can take care of the memory management issues for good. It is not that I can't do a VM, but I believe that guy when he says that making it would be hard work. I mean Spiral took me a year just to get it to this point as simple as it is. And I am not sure I can stomach the double whammy of leaving .NET and having to build up everything from scratch all over again.

The incentive for that kind of thing would have to be significant, GC on its own would not be enough.

Well, hopefully by that point I'll have made a killing using my AI methods and will have the leisure to make the right choice.

Right now my hands are tied and i can't spare too much effort for non-essentials.

Spiral was necessary because otherwise I could not guarantee the correctness of the Cuda kernels, and likewise, it is not like I am dealing with C here where writing past the ends of an array would overwrite its offset and explode the entire pogram down the line.

At worst I will either get leaks, or I will try to access a null and I can deal with that somehow.