As it turns out it was the same bug as last time.
All the tests finish in 7.75s. Yeah, this is it. I did it. I am done with the redesign of layout types.

Thankfully this error was easy to track down thanks to the effort I put in last time.

Now I'll take a break and when I get back I will start work on the new allocator. This will be the last time I'll get distracted. After I am done with this, I will finish the library and move to bigger and better things.