I am obsessing about the wavefront. Really, there isn't a simple way of doing it. I guess it is impossible to generalize loops fully. There are various ways of doing it, but each of them comes with its own complexity. It is impossible to do it without significantly bumping complexity up and wavefront iteration is pretty complex. Maybe with cooperative groups there might be a way around it?
Regardless, since it is so nontrivial I'll keep it for the last.

Let me start work on the regions. I do not feel like it, but I need to put in some work for the day.

Seriously, fuck the wavefront. The complicated compilation of fast NNs is really making things quite hard for me.

I guess one of the problems is that I still haven't decided how much flexibility is needed in a library.

Mastering the compilation of NNs does look like it will be lifetime of work.