Here are the streams and I even implemented waiting on them for good measure.
There is also the async thing, but that should wait. That actually needs to be a part of the Learning module. It does not make sense to implement it here because I have nothing to pass the stream to as an argument.

I am not sure if I feel like doing anything else for the day. This was really quite tedious.

Well, as lazy as I feel those 200 lines that have been written in the last two days had to have come from somewhere.

Tomorrow I will start work on the tensor and move to the main library after.

Hopefully it won't take to long.