Usually at this point I'd be programming for 2 hours, but today I broke that habit by reading a manga that was posted on /a/.
So my morning was complete indulgence. Every day is usually a fight to keep myself working, but yesterday when I admitted that the work I was doing was tedious I ended up cursing myself.

Sigh. The problem was that the last few hours since I got up were really great. It is just so pitiful that after the day is over and my energy has been spent, I can't enjoy this stuff to the same degree.

The whole reason why I started on this path was so that I can firmly grasp that feeling, so it is just sad that the path towards that goal leads in the opposite direction.

I do hope those of the future will never have to go through this. Life should be led by going closer to one's goal. It should not be a twisted version of that.

That having said, what I am doing really is tedious. The reason for that is that I am seemingly making an endless series of tradeoffs that are specific to the situation rather than coming to a firm solution. Work on the library has been like that for a while and it cleaves away my spirit.

The thing facing me right now is deciding how I want to design the tensor module. Do I want to leave the environment at the top of the module or do I want to weave it through the individual functions?

Agggghhhhh...

I'll spend some time agonizing over it today.