I refactored the map_redo into map_redo_map and had it stop transfering to host. In lieu of that I am just doing another reduce on the GPU instead.
It is better this. The extra step of using lazy values on lieu of that is not worth the complications to the library.

But I am going to have to redesign parts of the backwards pass in order to support this kind of thing.

I'll also need to redesign the environment so it supports switching regions and streams.

Well, forget that for a bit.

Let me get the basics done. I'll start from the first Learning test and go from there.