I've made the necessary breakthrough. I am going to go full out with OO for the library. Simply put, Spiral is not Haskell and the compiler is not helping me do all that refactoring. I need a different way of doing things.
I am going to have to go through all the tests from the top. I actually really really like the OO way actually. It solves pretty much all the problems that I have with the library currently and as a design pattern it is far superior to the functional one.

The issue is just that my experience with OO is too clouded by languages like C++ and C#, so I hadn't been able to attune to it. Spiral is just so more powerful so I can in fact emulate the OO of dynamic languages like Ruby and Smalltalk apart from interface mutation.

I should have gone with OO from the start, but this is just how my lack of experience in Spiral tends to manifest. I was absoutely right when I said two weeks ago that I am not yet at the point where I can draw out Spiral's true power.

Let me take a break here.

I will admit that there was some tedium up to now in doing this and a major part of it is because I was not exactly sure what the new design should be.

With this I've made the necessary conceptual leap and will be able to progress much more smoothly from here on out.