it is not perfect, but I think this should meet my requirements for the new object system. There are still pieces missing, but the new CudaTensor should serve well as a sketch.
A great benefit of this is that inside the CudaTensor module, I've completely broken the dependency on having to have context and the Cuda module passed in directly. Now that is quite something.

As usual at this time, I need to stop here for a bit, but when I get back I'll se about moving to testing.

Edit: Forgot to commit the above before I took off. At any rate, I added module_add and module_remove to the language.

Let me go forward. This tedium and the endless refactoring will end when the task is completed properly.