8:35am. I am up. What did I say about the social site ban?
Enough. Today I will dedicate whole heartedly to programming.

Somehow I've been thinking about trading during the night again.

The answer is magic. Just optimize the shit of it. I can see the truth now.

Suppose you have an agent of a certain level. Even if he can't push out the traders in the big stocks it might be possible to make good money bullying the ones in the weaker stocks.

The key is being able to create good agents.

Once that is possible to do reliably (which isn't right now in RL) then it also becomes possible to reason.

Instead of bashing your head against the wall by futilely trying to get better, it is possible to start from a position of strength. From that position it might be possible to infer that trading on stocks on which there is no profitability is futile - it is not a matter of getting better because you are already good. Then you can throw those away and search for profitable scenarios only.

With skill at abstraction it becomes possible to abstract away skill. That is the key; it is the long sought for holy grail.

8:45pm. I will not speculate again; solving the meaning of life and discovering the true self improvement method is enough. There is simply not beating that and there is no point in gazing into the abyss anymore.

But life is a series of games. To beat normals I have to get good at beating those. Three decades of my life have shown me that my ability of this department is lacking. The annoying thing is that in any field the top gets dug in and it is very hard to dislodge.

To have any chance at consistency I will simply have to get good at learning. Improving one's learning is impossible for humans, but it sure is possible for machines.

I am going to outright blast them from their holes. If they are anything less than perfect then they will die.

8:50am. Now it is time to start.

It is time to show my faith in my new principles. Abstract away skill. Optimize for power. Unify the shattered pieces the mind. Never lose again.

With every step, I get closer to my goal and today is going to be another one of those, all the way until I get there.

Really, I spent the last two weeks being incompetent. Let me finally enjoy the wonders of object oriented programming as strange as that sounds.

Edit: I forgot to push this again. I did a bit of work today

Is there anything I am missing. I do not think so. Let me proceed from here.

I really like the new architecture. It solves my extensibility problems quite neatly.

Let me move to testing.