The blas and the random tests work now.
With this things are slowly starting to fall into place. Let me take a break here and I will start work on the kernels after that. It won't take long.

After that it will be time for the real deal.

Now that I have library in this form it will not be hard to add concurency capabilities to Learning part of it.