...
I am completely stuck here and I am currently in full brainstorming mode. This happens to me when not only do I not know what the problem is, but I do not know how to debug it either.

Really, the accuracy working correctly when the cost diverges so badly, just what could it possibly mean?

Let me try a thing and then I will get away from the screen for a bit. For these kinds of bugs there have been cases where I failed to resolve them at all and had to resort to a complete redesign.

Memory safety is no joke.