"5:15pm. I am at my end here. I have no idea which commit should I go to. And I am not that thrilled at using this kind of proof technique.
Should I implement a standard allocator instead that uses Cuda's malloc natively?

5:20pm. No this won't make me happy. I want to reason through this. I want to get the success that comes from having superior skill. I want to make the allocator that I can prove to be absolutely correct."

I refuse to give up here. I will resolve this even if I have to break my brain in the process.

I am much better than I was before so I do not have to do something like the lame technique of  comparing pointers with the previous library.

Agh, this fucking thing, there is this thing that I just can't pry out. It is like a memory that refuses to bubble to the surface.

The truth is, the pointers do feel correct to me. Just what else could be wrong here?