inl gridDim = 1 //lit_min 64 (divup span blockDim)
When I do the above in map_redo_map it started working. Yeah, got you piece of crap.

The clue was the accuracy after all - it was an indication that only the cost is wrong. I could have done this 3 hours ago when I first became suspicious of it, but I could not justify such a move.

Still why doesn't it work?

I don't think that there would be an out of bounds error here. This is why I haven't tried a thing like this hours ago.