"6:35pm.
```
/// Flattens the tensor to 1d, maps, reduces it and maps it.
/// Map is optional. Allocates a temporary tensor for the intermediary results.
inl map_redo_map w {d with redo neutral_elem} (!zip (!flatten in)) =
    inl to_dev_tensor = w.CudaTensor.to_dev_tensor
    inl run {map_out map_in blockDim gridDim} (!to_dev_tensor in) =
        inl in_a :: () = in.dim
        inl out = w.CudaTensor.create {elem_type=type map_in in.elem_type |> map_out; dim=gridDim}
        inl out' = to_dev_tensor out

        w.run {
            blockDim gridDim
            kernel = cuda
                inl x =
                    grid_for {blockDim gridDim} .x in_a {state=dyn neutral_elem; body=inl {state i} -> redo state (map_in (in i .get)) }
                    |> cub_block_reduce {blockDim redo} |> map_out
                if threadIdx.x = 0 then out' blockIdx.x .set x
            }

        out

    inl map_in = match d with {map_in} -> map_in | _ -> id
    inl map_out = match d with {map_out} -> map_out | _ -> id

    inl in_a :: () = in.dim
    inl span = s in_a
    inl blockDim = lit_min span 256
    inl gridDim = lit_min 64 (divup span blockDim)

    if gridDim = 1 then
        run {map_out map_in blockDim gridDim} in
    else
        run {map_out=id; map_in blockDim gridDim} in
        |> run {map_out map_in=id; blockDim=gridDim; gridDim=1}
    <| 0
```

The real mistake was where I put run. If I did it in the beginning like this then I would have just gotten a variable out of bounds error.

This trivial thing just cost me a whole day. I did get a feeling the original was loose somehow, but it was just that - a feeling. I did not get the idea how to fix it even though it should have been obvious.

6:40pm. Yeah, now the test works correctly.

As much as I'd like to continue I think I will stop here for the day.

In the end I did what I wanted in the way I wanted it to. I fixed the allocator and then I fixed the bug that came out once I fixed it. I am not bad at debugging it seems. The old me would have been confused by the tag teaming of bugs, but this me knew that once I tightened the allocator that it would be correct.

Tomorrow I will do what I started today which is clean up the `run` function and make the CudaTensor more extensible, though maybe I'll just do the first one."

Well that is it for the day.

I really wish there was another manga like "Sentou Hakai Gakuen Dangerous" waiting for me when the day is over, but maybe I'll watch some HxH instead. I never went through the anime itself and most of the manga has already faded from memory so it is like I am discovering the series for the first time.

I like it, but one of these days I should really just find a game to play for when work is done.

Forget that. There will be enough time for play when the computer starts taking over my work. Really, the era when most of the computation is taken up by the GPU instead of my brain can't come soon enough.