I am going to take a few hours away from the screen to answer a very simple question - do I have enough to finally start on the RNNs. These last 3 weeks have worn down my spirit and I need to recenter myself.
I would have done this yesterday, but those unexpected bugs threw me off track.

If it is possible, then I will gather all of my power and finish the thing before the month is over.

I have to believe. I know that seems unlikely given how much time I spent not moving closer to my goal in the last 3 weeks, but once the pieces are set, the flame will spread quickly.

So it is with programming.