"11:55am. I am slowly coming up with an idea. I am going to take a break for a few hours away from the screen just to answer one very simple question - do I have enough to finally start work on the RNNs?
12:10pm. Let me go do the chores and then I will have some time away from the screen.

5:55pm. I spent most of this time in bed thinking and putting my effort into aligning my goals.

I have so many burdens right now and so many regrets. There are 3 problems, 3 phases and 3 goals. Programming, AI and money.

1) Programming - This is such a massive burden to me. I am really trying my best here, but I am at my end trying to get past where I am. Sure, I have a lot to go before I master Spiral, but looking at it from another angle Spiral is my limit. It is a tragedy, but I cannot imagine a language better than Spiral. Not within real world constraints facing me. A language better than it would pretty much have to be magic.

So, in this respect, my programming skill has reached it's limit and the most I will be able to do is refine it.

I hate it. Spiral itself could be so much better if I could put in ten years of effort just improving it instead of a single one, but I just do not have that kind of time or the will to undertake such an endeavor.

And it is embarrassing to me because I want to do better. I am ashamed, but the one to come will have to follow up on what I could not manage with my meager human abilities.

2) AI - Spiral was made so I can increase my programming skills which I can translate into skills in the AI domain. But it is such a great burden. I beat C in a very real sense. I made the language that the others could not do in four decades. I crystallized the staged functional programming style in Spiral.

And yet, much like when I discovered the truth of the self improvement process, all it did is uncover a huge obstacle in front of me.

https://www.alexirpan.com/2018/02/14/rl-hard.html

I mean look at this shit. This is what is waiting for me. I mean I predicted it would be like this years ago, but that does not mean it will be any less work.

What I did in Spiral is just the beginning - it is not enough to significantly extend the state of the art in a single field. I am going to have repeat the performance. I am going to have to crystallize the art of differentiable programming and replace those gimmicky RL algorithms with things that work well.

3) Money - This thing seems always out of my reach. It'd be a lot easier on me if I could get intermittent cash rewards for playing this game rather than the ever distant big pile at the end. This goal is also a great burden on me. I just want to skip #2 and do this one first, much like taking a loan, and then I'd pay it off by doing #2.

But things do not work that way.

No way will I magically overcome all the obstacles in #2 and go on my happy money making shadow online killing spree.

I cannot realistically take responsibility for #3 much like I cannot for #1 anymore.

Just like I cannot compress programs beyond a certain point in terms of reasoning even with Spiral, so I cannot deal with this goal.

I can either focus on #2 or #3. And since going for #3 is just placing an unnecessary burden on my, I will ditch it. Hence, until I've put in a significant amount of time, possibly two years into #2, I cannot have money as a measure of my success.

Fundamentally, I have to face the fact that even though I prepared 3 years just to have a go at this, I might not succeed in attaining the AI skills necessary to meet my other goals. I might stumble at #2 or some disaster might happen that would end the quest.

So I have to focus and I have to make my own rewards.

In terms of programming skill, I am in my own league.

In terms of AI skill, I am a baby.

So from now on, I will forget all that is auxiliary. I will make pumping up my AI skill my main goal with nothing else as a priority. From now on my worth as a person will be that measure and I will place weight of my life on that scale.

By doing so I will be able to shed my burdens and for once stop seeing what is to come as an immeasurable mountain of work before I can get to #3. I will be able to able to play the game as it should be played.

I won't be too bad.

I've been having difficulty moving forward for the past few days, but the library itself is mostly finished even the recurrent segments. I just need to strip them from my mind and translate them into code. It won't be a problem.

The ML library is not a problem anymore.

It is time to move into the world beyond it. On my will and on my desire, in front of me is not a huge towering wall, but a door. I intend to pass through.

7:05pm. Writing the above was important to me because I've been spending so much mental energy trying to futilely compress things in the ML library. It is important to know when the job is over. One more step and it will be.

It is important to know when to change gears and the mindset.

It is important to crystallize and focus it and tap into it.

Starting tomorrow I won't think of myself as language designer, but an AI student."