I am up. I am still groggy.
I've been gathering my power throughout the past day and I do feel more motivated today. I've been feeling pretty squeezed trying to keep the size of the library from exploding, but really there is only so much I can do here.

As I go into reinforcement learning I am going to have to write code and a lot of it. It cannot be avoided that by the end of the year, the library will be a lot larger than it is now. So forget it. Embrace it. The programing practice is now over. It is time that I get the functionality.

For the sake of leveling up my AI skills.

The most important thing is to focus on the goal and to draw strength from it. The place I am going should not be a chore.

And money aside, AI skills on their own are a treasure. They are more than worth being a goal in itself. I do not have to have money to justify them. I need to remember the feeling when I first started. My vision.

Just like it was back then, I can't really see the current crop of RL algorithms as anything but a gimmick unlike the supervised deep learning stuff.

I think that RL at the current juncture meshes particularly poorly with RL, so my basic idea was that I should follow basic engineering principles and see what comes out of that.

Forget all the disparate shattered pieces of the library in my mind. Just go forward. They will fall into place as things move along.

Once again, I will make it my goal to make a recurrent network. I wanted to do this three weeks ago, but I ran into difficulties with resource management. Now the resources are threaded throughout the library and I have nothing more to ask for.

Press onward.