This is it. I do not have to append the cost directly. All I have to do is run the things in tandem.
I am so engrossed into this. This was amazing.

I pretty much solved everything here and set up the stage for the delayed runners.

I need to have breakfast here. I am going to have to take the rest of the day off due to other obligations, but I'll get some words in later what I am aiming for here.