I am done for the day. I did some really good design work that will pay for itself many times over. Honestly, it did not occur to me that for recurrent layers I should just use the facade pattern.
But this is good, really good as it simplifies things greatly. I will be able to fit the wavefront iteration into it somehow.

Let me take a break here for the day.

Also unfortunately, Github is really ferocious at  referencing commits and I did not want that huge essay to end up in the first issue. Let me figure out if I can ammend the past commit.