Just like I've said a few days ago, the time has come to put my obsession regarding programming behind me.
In recognition of this fact, I will start with the wavefront.

I've been obsessed about it for months now and it itself is a major source of burden to me. The way I wanted to do it could be done in theory, but it wouldn't be worth it in practice.

I am just trying to satisfy my own vanity.

In truth, the correct here is to do it normally, but chain the streams together so it comes out to the same thing.

The new HyperQ schedulers in the Kepler and onward Nvidia cards are the solution to my problems here.

I am just so weak. I have Spiral, but its powers hardly make me all powerful. The rules haven't changed - every piece of functionality has a cost associated with it.

I am dependent on what the hardware allows me to do. I am dependent on what the .NET VM allows me to do.

I cannot avoid paying those costs.

The non-contiguous wavefront iterator I've been trying to materialize is just not realistic - it is far to expensive as a feature compared to simply chaining the streams and letting the GPU scheduler take care of the rest.

Ah damnit, everything is solved. That is how I feel like right now.

...It is annoying, but with these pieces and with the unrealistic wavefront and fully fused passes put behind me - if I assume that the combinatorial NN library design I am converging to right now is optimal, then there are no more obstacles to me finishing it.

I will be done in less than a month. There simply isn't that much left to solve.

I do not feel like starting right now because I feel obsessed. Let me get away from the screen for a bit to gather my power again and then I will let loose again. I need to refocus and really put those unrealistic urges behind me. Once more, I need to focus until I've internalized the new perspective completely.