I am still thinking about this, but unrolling that state for only a specified number of steps really blows.
I'll make a specialized loop for this sort of thing tomorrow.

I did enough for today. As usual progress is going at a crawl, but I will at least deal with the state changing loops tomorrow.

After that I will be able to do the iterators for the recurrent nets.

Figuring out how to make the transition function is interesting and somewhat annoying at the same time.

It really is a great challenge of programming skill.

By the end of the month I will have gone through it. I will finish the char RNN.