Today I am in action from the very first moment.
I did some redesigning. Let me finally implement unroll, and then I will figure out how to deal with the unrolling loops.

This work is quite hard and intricate. If I can get it right, it will give me a major advantage.

When it is done, the actual Learning part of the Spiral library won't be too large, but it will be quite effective, more so that in any other library.