I am a lot more satisfied with these few changes.
Now fold does all the work and iter is the one that iterates over the dataset.

This is ideal. It is not so bad.

I'll make the foldl for ResizeArray tomorrow.

Let me finally stop for the day. Tomorrow I might have enough to finally move to testing. I do not have to worry about bringing in softmax and everything else just yet.

I still have some clearning left to do.