I've autistically moving terms around for hours now trying to capture the essence of those functions. Without a doubt I have it now. Well, I am missing checking for nans, but the essence of the function is neatly captured within. There is no beating this.
I was really unhappy how in the previous version the damn iterator did so many things apart from run the net and up to now I actually did not have an idea how to solve it.

What I had was essentially inherited from the old F# library.

As it turns out the solution to everything are functions returning functions and the facade pattern.

Now, I really need to do the iterators for the feedforward case. I was supposed to be working on it, but then I started working the recurrent ones again.

Ahhh, let me just get on with it. It shouldn't be too hard now that I've caught scent of it.