This is absolute perfection. This is what the iterators should be and should been from the start. They do only the most essential work for them and as such attain the maximum utility. I'll add short circuiting at a later time.
Regarding the wavefront, it would be better if it was all async on the host as well. If it was I'd be able to guarantee memory safety, but Cuda doing things half way is the usual for it. I won't bother myself over it anymore.

Let me commit here.

I think I will take a break here.