"6:45pm. Now I am just getting lost in thought with regards to initializers. If only I could merge passing of inputs with the rest of the passes...but it should be doable.
6:50pm. Yeah, forget this shitty init. I already tried the next layer thing and for some reason I came to the conclusion that it would not work.

But it actually would definitely. For gathering weights I could just use the damn monad.

6:55pm. But I'd have to change the layer design and it would come with its own weaknesses.

Well, forget it. As I said, if I want different behavior I am going to have to write a different library. This one is designed for straightforward layer level architectures.

6:55pm. This is the twilight of programming, when the limit is reached and there are only tradeoffs.

Spiral will allow me to ride it out. I have to believe in it.

Honestly, this library is not such a big deal. I just wasted a lot of time going off course in the last two months. I will finish the rest of the tests tomorrow and hopefully finish the char RNN in two days."

Yeah, two days should be a reasonable estimate for getting the first version out. I'll get quite a lot done by the time the month is up.