"6:55pm. This is the twilight of programming, when the limit is reached and there are only tradeoffs.
Spiral will allow me to ride it out. I have to believe in it.

Honestly, this library is not such a big deal. I just wasted a lot of time going off course in the last two months. I will finish the rest of the tests tomorrow and hopefully finish the char RNN in two days.

7pm. Ah, no wait, I get it.

It should be definitely possible to make binary layer combinators. While right now I could not chain layers as they are unary, I could make a function that takes two layers and pipes between them. I could also take care of the `init` in that place. Yeah, that should be it.

...Ah, I see it. Yeah, I've solved it. There are no tradeoffs here at all. I can get this whenever I want.

7:10pm. No wait no, it is not like that. It is incomplete.

7:30pm. I slack when I need to work and now I work when I need to slack. How hopeless.

8:20pm. I picked a late time to lie down and get up.

But this issue has me obsessed.

Basically, I get it now. This actually has a lot to do with types. It seems amazing that I can chain operations in arbitrary way, but as soon as they become stateful, it becomes absolutely impossible to do anything.

This reminds me of the situation in the old languages where all the variables and their types needed to be declared up front right at the beginning of the function. I am thinking of Pascal and Fortran and the old C.

8:40pm. No actually, it should be possible. I am focusing like crazy on this right now.

I know for a fact that given an initialized layer passed into another array, I can get the whole network like that.

And I know that given a result passed to a function, I can get the whole thing. So layers are like types and results are like values.

There should be a way to make them happen in parallel.

9:40pm. My hatred has shown me the truth. What I did in the past few hours was instead of resting, I went into an extremely deep trance.

I figured out everything. Every fucking thing.

I know exactly how to deal with layers with zero tradeoffs and with maximum expressiveness and code reusability.

This idea occurred to me a few times before, but only now can I see it in every little detail and I can understand that it is in fact the best choice.

I'll go into more detail tomorrow."

This is what I do all day. I brainstorm.

This experience of having to deal with layers is really making me appreciate why languages in languages are necessary.

I am a madman, but I am going to overcome this. Without a doubt, I will finish the library and move from there. Once I do it perfectly, the rest will fall away.

But if I do it incorrectly, I will just get tied down in maintenance.