I've been doing a lot of thinking. A lot.
The graph approach is basically what I need here. Simply put, I need a small language just to have the comfort of being able to definite and initialize layers properly. The way I am doing it right now is nonsense.

What I will aim for layers is to make a compile time interpreter over a graph. So far I've been doing the exactly the reverse and it is simply too difficult.

If the layers weren't stateful it would be no problem, but the approaches I've been trying to take are impossible.

They do have many great elements though - the facade pattern in the layers was still a thing of genius and I am definitely going to be reusing it.

Taking a symbolic graph approach will have many benefits down the line such as allowing me to checkpoint the network. And loading and saving would be very difficult using my current approach.

I've been right to feel dissatisfaction.

The feeling I get here is the same as during my work on Spiral in the early days where one day I would decide to do a thing and the next day I would change my mind on the design.

Golden days.

These are my golden days of ascent. They are painful and tedious, but I will not look at the time I've spent here with regret when they are over.

Well, sentimentality aside, I am going to have to redo the library again. When I was fresher I used to think all this refactoring is an aberration, but it is in fact programming itself.

In the end, the task is done only when it is done perfectly.